% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_randomBackground.R
\name{helper_randomBackground}
\alias{helper_randomBackground}
\title{Randomly sample background regions}
\usage{
helper_randomBackground(pool, focal, n_ratio, bgReplace = FALSE, seed = NULL)
}
\arguments{
\item{pool}{A GRanges of candidate bg regions.}

\item{focal}{A GRanges of foreground regions.}

\item{n_ratio}{Numeric; number of bg = n_ratio * number of fg.}

\item{seed}{Integer; passed to set.seed() for reproducibility (default NULL)}
}
\value{
A GRanges of randomly sampled background regions.
}
\description{
Randomly sample background regions
}
