% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_resolveBSgenome.R
\name{helper_resolveBSgenome}
\alias{helper_resolveBSgenome}
\title{Resolve and load a BSgenome for a given UCSC build}
\usage{
helper_resolveBSgenome(organism, genomeBuild, bsMasking = FALSE, suffix = NULL)
}
\arguments{
\item{genomeBuild}{Character(1). UCSC‐style genome build (e.g. \code{"hg38"}).}

\item{bsMasking}{Logical(1). Whether to require the “masked” version
(i.e. masked repeats/ambiguous regions) (default \code{FALSE}).}

\item{suffix}{Optional character. A package name suffix (e.g. \code{"dbSNP151.major"}),
to disambiguate multiple hg38 variants.}
}
\value{
A loaded \pkg{BSgenome} object.
}
\description{
Looks first among installed BSgenome data packages, then falls back to
Bioconductor’s available genomes list (and installs if needed).
}
\examples{
\dontrun{
  # load your installed hg38 BSgenome or install it if needed
  bs <- helper_resolveBSgenome("hg38")
  # get the masked version if you prefer
  bs_masked <- helper_resolveBSgenome("hg38", bsMasking = TRUE)
  # or explicitly ask for dbSNP151.major
  bs_major <- helper_resolveBSgenome("hg38", suffix = "dbSNP151.major")
}
}
